fix #56 highlighting of topnav section when in subpage --
Whenever page.url contains item.url the 'active' class is applied to the list item of the topnav. As
a result the 'active' class is lost when navigating to a sub page in the current section.
To solve this issue we exploit the association of sections of the website with
sidebars, as scoped in _config.yml.

To do so we introduce a new key 'section' in topnav.yml to formalise the notion of a website section.
We then check the current sidebar against the one expected by the current section and
keep the 'active' class when they match.