Make native doxygen website more visible (#106)
This adds an additional section referring to the native C++ API and moves the Doxygen link there. This should make the API documentation more visible and easier to find as not everyone might know Doxygen. Before, the link to the API documentation was hidden in the running text. The change also makes the representation of available APIs more consistent as all APIs are represented by a tables now.

Co-authored-by: Benjamin Uekermann <benjamin.uekermann@gmail.com>