Update adapter-calculix-get-calculix.md
At the time of writing this, {{site.calculix_version}} points to 2.2 which is wrong and makes the links broken. If it points to 2.20 then there will be no problem. For the time being I change {{site.calculix_version}} to 2.20 which is not the smartest thing to do but solves the issue anyway. 
Also there is typo in "Summary". I changed "challenges" to "challenge".