Document coupling to remote systems
@LouieVoit published a [very interesting paper](https://doi.org/10.1007/978-3-031-40843-4_42), using preCICE to couple a CFD solver running on a laptop (_"while waiting at the airport and connected to a poor wi-fi"_) to a massively-parallel MD simulation running on a cluster.

The paper is helpful to demonstrate the concept, but lacks some technical details. I got these from Louis via e-mail and I am summarizing them in this documentation addition at the bottom of the [communication configuration](https://precice.org/configuration-communication.html) page.

Note that this is not the first user that has asked for something like this. In the context of the helicopter simulations Master's thesis I supervised, we had a very similar issue, where each code was licensed for a different system.

- @LouieVoit Have I forgot any details?
- @fsimonis Does this make sense?
- @davidscn Do you maybe want to test this? Could be very easy for you, since you daily work on a cluster.